Refactor some code around
